requirements.txt
torch
opencv-python
pillow
diffusers
transformers
insightface
scipy
accelerate
safetensors
